+++ releases/eucalyptus/3/wb/config/lms/docker_run_production.py
@@ -33,7 +33,7 @@ from xmodule.modulestore.modulestore_settings import (
 )
 
 from ..common import *
-from .utils import Configuration, ensure_directory_exists, prefer_fun_video
+from .utils import Configuration, prefer_fun_video
 
 # Load custom configuration parameters from yaml files
 config = Configuration(os.path.dirname(__file__))
@@ -927,10 +927,6 @@ ORA2_FILE_PREFIX = config("ORA2_FILE_PREFIX", default=ORA2_FILE_PREFIX)
 
 # If backend is "filesystem"
 ORA2_FILEUPLOAD_ROOT = DATA_DIR / "openassessment_submissions"
-# The code in edX ORA2 is missing appropriate checks so we must ensure here that this
-# directory exists:
-ensure_directory_exists(ORA2_FILEUPLOAD_ROOT)
-
 ORA2_FILEUPLOAD_CACHE_NAME = config(
     "ORA2_FILEUPLOAD_CACHE_NAME", default="openassessment_submissions"
 )
@@ -1395,22 +1391,15 @@ FUN_THUMBNAIL_OPTIONS = {}
 # -- Certificates
 CERTIFICATE_BASE_URL = MEDIA_URL + "attestations/"
 CERTIFICATES_DIRECTORY = MEDIA_ROOT / "certificates"
-# The code in fun-apps is missing appropriate checks so we must ensure here that this
-# directory exists:
-ensure_directory_exists(CERTIFICATES_DIRECTORY)
 
 STUDENT_NAME_FOR_TEST_CERTIFICATE = "Test User"
 
 # Videofront subtitles cache
-VIDEOFRONT_SUBTITLE_CACHE_ROOT = DATA_DIR / "video_subtitles_cache"
 CACHES["video_subtitles"] = {
     "BACKEND": "django.core.cache.backends.filebased.FileBasedCache",
     "KEY_PREFIX": "video_subtitles",
-    "LOCATION": VIDEOFRONT_SUBTITLE_CACHE_ROOT,
+    "LOCATION": DATA_DIR / "video_subtitles_cache",
 }
-# The code in edX ORA2 is missing appropriate checks so we must ensure here that this
-# directory exists:
-ensure_directory_exists(VIDEOFRONT_SUBTITLE_CACHE_ROOT)
 
 # Used by pure-pagination app,
 # https://github.com/jamespacileo/django-pure-pagination for information about